+++ src/V1/Screens/ProjectScreen/index.tsx
@@ -1,7 +1,11 @@
-import React, { useState, memo, useCallback } from 'react';
+import React, { useState, memo, useCallback, useMemo } from 'react';
 import { useLocalSearchParams } from 'expo-router';
 
+import SubscriptionManager from '@SubscriptionManager';
+
+import { translations } from '@V1/Translations/index';
 import { SampleSettings } from '@V1/Types/ProjectTypes';
+import ConfigService from '@V1/Services/ConfigService';
 import CacheService from '@V1/Services/CacheService';
 import AlertService from '@V1/Services/AlertService';
 
@@ -9,18 +13,28 @@ import { Animation } from '@V1/Animation/index';
 import { Layout } from '@V1/Layout/index';
 import { TC } from './__TC__';
 import { LC } from './__LC__';
+import { ErrorCodes } from '@V1/Globals/ErrorsCodes';
 
 export const ProjectScreen = memo(() => {
 
   const id_project = useLocalSearchParams().id_project as string;
+  const config     = useMemo(() => ConfigService.config, []);
+  const RError     = useMemo(() => translations.global.errors[config.language], []);
   const [samples, setSamples] = useState<SampleSettings[]>(CacheService.allSamples);
 
   const onCreateSample = useCallback(async () => {
-    await AlertService.handleAlert(true, {
-      type: 'sample creation',
-      id_project: id_project,
-    }, () => setSamples(CacheService.allSamples));
-  }, []);
+    if (SubscriptionManager.freeUserLimitCheck(samples.length >= 5)) {
+      await AlertService.handleAlert(true, {
+        type: 'Buy Subscription',
+        message: RError(ErrorCodes.FREE_USER_SAMPLE_CREATION_LIMIT),
+      }, () => {});
+    } else {
+      await AlertService.handleAlert(true, {
+        type: 'sample creation',
+        id_project: id_project,
+      }, () => setSamples(CacheService.allSamples));
+    }
+  }, [samples]);
 
   const onUploadProject = useCallback(async () => {
     await AlertService.handleAlert(true, {
@@ -33,8 +47,8 @@ export const ProjectScreen = memo(() => {
     <Layout.Screen
       screenButtons={
         <TC.ScreenButtons
-          onCreateSample={onCreateSample}
-          onUploadProject={onUploadProject}
+          onCreateSample={async () => await onCreateSample()}
+          onUploadProject={async () => await onUploadProject()}
         />
       }
     >