+++ docs/ReleaseNotes.md
@@ -1,5 +1,7 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 1.10.12 Animations updated
+# 1.10.13 Free user paywall and Subscription sku
 
-- Now animations use useAnimatedStyle methods to animate.
+- Changed the premium sku to monthly premium
+- Added a free user paywall condition on key points
+- IAP scope deleted. Instead now the app has a BuySubscription alert